refactor: divide popup.tsx into template and container
